Extract StDistanceFunction
